Change bpm field from int to float in sql db

+++ sql/db.sql
@@ -9,7 +9,7 @@ CREATE TABLE arrangements (
     id SERIAL PRIMARY KEY,
     user_id INT REFERENCES users(id) ON DELETE CASCADE,
     name VARCHAR(255) NOT NULL,
-    bpm INT NOT NULL,
+    bpm FLOAT NOT NULL,
     tags VARCHAR(1000) NOT NULL,
     file_name VARCHAR(255),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
